Powerset problem
If S is a set, then P(S), the 'powerset' of S is the set of all subsets of S including the empty set and S itself. If S has cardinality N then the cardinality of P(S) is 2N.
Write the function 'powerset' that computes the powerset of a given set. Your function will satisfy this signature:
val powerset : 'a list -> 'a list list
For example, in the top-level:
# powerset [1; 2; 3] ;;
- : int list list = [[]; [3]; [2]; [2; 3]; [1]; [1; 3]; [1; 2]; [1; 2; 3]]
The ordering of the sets is not important here, just that all the sets are discovered.