Sieve of Eratosthenes
The sieve of Eratosthenes is a simple, beautiful algorithm from antiquity for finding all prime numbers up to any limit. It works iteratively by marking as composites the multiples of each prime, starting with 2. It is in fact, one of the most efficient ways known to find all of the smaller primes.

For example, to find the prime numbers less than 10, start with the list [2; 3; 4; 5; 6; 7; 8; 9]. Removing multiples of 2 you get the list [2; 3; 5; 7; 9], then removing multiples of 3 you get [2; 3; 5; 7] and since there are no multiples of 5 or 7 for this limit, we are done.