Huffman encoding
The type α tree is suitable for modeling "binary trees".

type 'a tree = | Leaf of 'a | Node of 'a tree * 'a tree

We can associate a node or a leaf in a tree of type α tree with the list of binary choices to make in order to traverse the tree from the root to that node or leaf. For example, if we denote left by the letter l and right by the letter r, then in the tree  Node (Leaf 3, Node (Leaf 4, Leaf 5)), rl represents the path designating the subtree Leaf 4.

(1) Given a tree t and a path of directions u,  we will denote the subtree of t corresponding to u (if it exists) by t/u. Define a type direction with two values, L and R, together with a function of type direction list -> α tree -> α tree option that starting with a list of directions dl and a tree t, provides t/dl.

(2) If all the values in a tree of type α tree are distinct, then the tree provides a binary encoding of the values it contains. For example, if we let 0 indicate left and 1 indicate right then the tree in this figure

To decode such a group of binary values by a binary tree of type  α tree and to produce the group of values of type α to which they correspond, all we have to do is traverse the tree, interpreting the binary symbols as indications of direction and returning to the root of the tree after getting to a leaf. In our example, the group "00111" decodes as "bee".

Write a function of type direction list -> 'a tree -> 'a list option  to decode the binary groups of type direction list by means of a tree.

(3) For a set of values that we want to code as binary, there are a priori a great many possible trees corresponding to various encodings. We can compare different encodings by comparing the average length of encoded messages. The following technique due to Huffman, finds the code that is optimal from this point of view. We start from a set of symbols (x1, ... xn) to code, and we associate them with probabilities (p1, ..., pn) that they will appear. Then we build the list [Leaf x1, p1; ... ; Leaf xn, pn]. Then from this list, we iterate the construction that consists of choosing (from the list) the two pairs (a, pa) and (b, pb) where pa and pb are the two smallest probabilities and replacing them with the pair (Node (a, b), pa + pb). When the list contains only a single pair, its first component is an optimal code for the original set. Write a function of type (α * float) list -> α tree that implements this algorithm.

This case,
# let tree = huffman ['a', 0.4; 'b', 0.35; 'c', 0.2; 'd', 0.05];;
val tree : char tree =
  Node (Leaf 'a', Node (Node (Leaf 'd', Leaf 'c'), Leaf 'b'))

corresponds to the encoding
[('b', 11);('c', 101);('d', 100);('a', 0)]
 
But note, the encoding is not unique. The step in the algorithm whereby you merge two nodes together does not mandate a choice as to which of the merged nodes becomes the left child and which the right in the aggregate.