Max contiguous sum
This exercise apparently has been a Facebook interview question. It goes like this. You are given an array of n integers, each of which may be positive, negative or zero. Give an algorithm to identify the start and end index, i and j say with j ≥ i, for (each of) the interval(s) whose elements form the maximal sum of all possible intervals. For example given the sequence [1; 3; -8; 2; -1; 10; -2; 1], the sum over the interval [3, 5], which has value 11, is maximal.