Breadth first graph traversal
I found it difficult to design my own data types and implement simple graph traversal, but it was a lot of fun to read back through graph theory. We also got to learn a lot about using records and external OCaml modules. It was probably the most fun and in-depth exercise so far.