Simple compression of arrays
Implement the so-called "runtime-length encoding data compression method". This is a function that packs consecutive elements of a list into sublists. For example, given let x = pack ['a';'a';'a';'a';'b';'c';'c';'a';'a';'d';'e';'e';'e';'e'] we expect x to have the value, [['a';'a';'a';'a'];['b'];['c';'c'];['a';'a'];['d'];['e';'e';'e';'e']] Next, encode the result of pack into a list of pairs (n, e) where n is the number of occurrences of the expression e. For example, given let y = encode (pack ['a';'a';'a';'a';'b';'c';'c';'a';'a';'d';'e';'e';'e';'e']) then we expect y to have the value [(4, 'a'); (1, 'b'); (2, 'c'); (2, 'a'), (1, d); (4, 'e')].